refactor: moved ServiceType / ServiceId in sub package
