+++ database/src/main/kotlin/dev/redicloud/database/repository/DatabaseRepository.kt
@@ -1,6 +1,7 @@
 package dev.redicloud.database.repository
 
 import dev.redicloud.database.DatabaseConnection
+import dev.redicloud.logging.LogManager
 import org.redisson.client.codec.BaseCodec
 
 abstract class DatabaseRepository<T>(val connection: DatabaseConnection, val name: String, val codec: BaseCodec? = null) {
