+++ utils/src/main/kotlin/dev/redicloud/utils/service/ServiceId.kt
@@ -1,4 +1,4 @@
-package dev.redicloud.utils
+package dev.redicloud.utils.service
 
 import java.util.UUID
